requirements.txt
apache-airflow==2.2.4
pandas==1.4.0
numpy
tabula-py
sqlalchemy
streamlit
pytest
pytest-mock
matplotlib
plotly==5.6.0
scipy
black